Replace 'recipes' by 'samples' in package and artefacts
