+++ lightbend-telemetry/lightbend-telemetry-java-mvn/hello-api/pom.xml
@@ -4,7 +4,7 @@
                       http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
     <parent>
-        <groupId>com.lightbend.lagom.recipes</groupId>
+        <groupId>com.lightbend.lagom.samples</groupId>
         <artifactId>lightbend-telemetry-java-mvn</artifactId>
         <version>1.0-SNAPSHOT</version>
     </parent>
