+++ lightbend-telemetry/lightbend-telemetry-java-mvn/hello-impl/src/main/resources/application.conf
@@ -1,4 +1,4 @@
-play.modules.enabled += com.lightbend.lagom.recipes.cinnamon.hello.impl.HelloModule
+play.modules.enabled += com.lightbend.lagom.samples.cinnamon.hello.impl.HelloModule
 
 # Configuration below adapted from the Cinnamon example for Java and sbt:
 # https://developer.lightbend.com/docs/cinnamon/current/getting-started/lagom_java.html
