+++ lightbend-telemetry/lightbend-telemetry-java-mvn/pom.xml
@@ -3,7 +3,7 @@
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0
                       http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
-    <groupId>com.lightbend.lagom.recipes</groupId>
+    <groupId>com.lightbend.lagom.samples</groupId>
     <artifactId>lightbend-telemetry-java-mvn</artifactId>
     <version>1.0-SNAPSHOT</version>
