+++ mixed-persistence/mixed-persistence-java-sbt/build.sbt
@@ -1,4 +1,4 @@
-organization in ThisBuild := "com.lightbend.lagom.recipes"
+organization in ThisBuild := "com.lightbend.lagom.samples"
 version in ThisBuild := "1.0-SNAPSHOT"
 
 // the Scala version that will be used for cross-compiled libraries
