--- mixed-persistence/mixed-persistence-java-sbt/hello-api/src/main/java/com/lightbend/lagom/recipes/mixedpersistence/hello/api/GreetingMessage.java
@@ -1,9 +1,0 @@
-package com.lightbend.lagom.recipes.mixedpersistence.hello.api;
-
-import lombok.NonNull;
-import lombok.Value;
-
-@Value
-public final class GreetingMessage {
-    @NonNull String message;
-}
