--- mixed-persistence/mixed-persistence-java-sbt/hello-api/src/main/java/com/lightbend/lagom/recipes/mixedpersistence/hello/api/UserGreeting.java
@@ -1,10 +1,0 @@
-package com.lightbend.lagom.recipes.mixedpersistence.hello.api;
-
-import lombok.NonNull;
-import lombok.Value;
-
-@Value
-public class UserGreeting {
-    @NonNull String id;
-    @NonNull String message;
-}
