+++ mixed-persistence/mixed-persistence-java-sbt/hello-api/src/main/java/com/lightbend/lagom/samples/mixedpersistence/hello/api/GreetingMessage.java
@@ -1,0 +1,9 @@
+package com.lightbend.lagom.samples.mixedpersistence.hello.api;
+
+import lombok.NonNull;
+import lombok.Value;
+
+@Value
+public final class GreetingMessage {
+    @NonNull String message;
+}
