--- mixed-persistence/mixed-persistence-java-sbt/hello-impl/src/main/java/com/lightbend/lagom/recipes/mixedpersistence/hello/impl/entity/HelloState.java
@@ -1,14 +1,0 @@
-package com.lightbend.lagom.recipes.mixedpersistence.hello.impl.entity;
-
-import com.lightbend.lagom.serialization.CompressedJsonable;
-import lombok.NonNull;
-import lombok.Value;
-
-/**
- * The state for the {@link HelloEntity} entity.
- */
-@Value
-public final class HelloState implements CompressedJsonable {
-    @NonNull String message;
-    @NonNull String timestamp;
-}
