+++ mixed-persistence/mixed-persistence-java-sbt/hello-impl/src/main/resources/application.conf
@@ -1,4 +1,4 @@
-play.modules.enabled += com.lightbend.lagom.recipes.mixedpersistence.hello.impl.HelloModule
+play.modules.enabled += com.lightbend.lagom.samples.mixedpersistence.hello.impl.HelloModule
 
 # Disable the Lagom JDBC persistence module.
 # This requires some Guice bindings to be explicitly set: see HelloModule.
