+++ mixed-persistence/mixed-persistence-scala-sbt/hello-impl/src/main/resources/application.conf
@@ -1,4 +1,4 @@
-play.application.loader = com.lightbend.lagom.recipes.hello.impl.HelloLoader
+play.application.loader = com.lightbend.lagom.samples.hello.impl.HelloLoader
 
 # Cassandra write-side configuration.
 hello.cassandra.keyspace = hello
