+++ mixed-persistence/mixed-persistence-scala-sbt/hello-impl/src/test/scala/com/lightbend/lagom/samples/hello/impl/HelloServiceSpec.scala
@@ -1,0 +1,38 @@
+package com.lightbend.lagom.samples.hello.impl
+
+import com.lightbend.lagom.scaladsl.server.LocalServiceLocator
+import com.lightbend.lagom.scaladsl.testkit.ServiceTest
+import org.scalatest.{AsyncWordSpec, BeforeAndAfterAll, Matchers}
+import com.lightbend.lagom.samples.hello.api._
+
+class HelloServiceSpec extends AsyncWordSpec with Matchers with BeforeAndAfterAll {
+
+  private val server = ServiceTest.startServer(
+    ServiceTest.defaultSetup
+      .withCassandra(true)
+  ) { ctx =>
+    new HelloApplication(ctx) with LocalServiceLocator
+  }
+
+  val client = server.serviceClient.implement[HelloService]
+
+  override protected def afterAll() = server.stop()
+
+  "Hello service" should {
+
+    "say hello" in {
+      client.hello("Alice").invoke().map { answer =>
+        answer should ===("Hello, Alice!")
+      }
+    }
+
+    "allow responding with a custom message" in {
+      for {
+        _ <- client.useGreeting("Bob").invoke(GreetingMessage("Hi"))
+        answer <- client.hello("Bob").invoke()
+      } yield {
+        answer should ===("Hi, Bob!")
+      }
+    }
+  }
+}
